+++ src/libtorch_c/torch_extensions/torch_redis_value.h
@@ -13,31 +13,31 @@ struct RedisValue : torch::CustomClassHolder {
 
   public:
     RedisValue(RedisModuleCallReply *reply) {
-        if (RedisModule_CallReplyType(reply) == REDISMODULE_REPLY_ARRAY) {
-            size_t len = RedisModule_CallReplyLength(reply);
-            for (auto i = 0; i < len; ++i) {
-                RedisModuleCallReply *subReply = RedisModule_CallReplyArrayElement(reply, i);
-                RedisValue value(subReply);
-                // arrayValue.push_back(value);
-            }
-        }
+        // if (RedisModule_CallReplyType(reply) == REDISMODULE_REPLY_ARRAY) {
+        //     size_t len = RedisModule_CallReplyLength(reply);
+        //     for (auto i = 0; i < len; ++i) {
+        //         RedisModuleCallReply *subReply = RedisModule_CallReplyArrayElement(reply, i);
+        //         RedisValue value(subReply);
+        //         arrayValue.push_back(value);
+        //     }
+        //     return;
+        // }
 
         if (RedisModule_CallReplyType(reply) == REDISMODULE_REPLY_STRING ||
             RedisModule_CallReplyType(reply) == REDISMODULE_REPLY_ERROR) {
             size_t len;
             const char *replyStr = RedisModule_CallReplyStringPtr(reply, &len);
-            // PyObject *ret = PyUnicode_FromStringAndSize(replyStr, len);
-            // if (!ret) {
-            //     PyErr_Clear();
-            //     ret = PyByteArray_FromStringAndSize(replyStr, len);
-            // }
-            // return ret;
+            stringValue= replyStr;
+            return;
+
         }
 
         if (RedisModule_CallReplyType(reply) == REDISMODULE_REPLY_INTEGER) {
-            long long val = RedisModule_CallReplyInteger(reply);
-            // return PyLong_FromLongLong(val);
+            intValue = RedisModule_CallReplyInteger(reply);
+            return;
         }
+
+        throw(std::runtime_error("Unsupported redis type"));
     }
 
     virtual ~RedisValue() {
